requirements.txt
requests
pyyaml
python-dateutil
https://github.com/crowdsecurity/pycrowdsec/archive/main.zip